List group

+++ resources/js/Pages/Admin/Front/Sections/List.vue
@@ -1,47 +1,45 @@
 <template>
-    <EmptyList :show="!sections?.data?.length" :is-filter="isFiltering" />
+    <ListGroup :is-filtering="isFiltering" :total-items="sections.data.length"
+        :pagination-pages="sections?.meta?.links">
+        <div class="row justify-content-center">
+            <ListItem v-for="section in sections?.data" :key="section.id" :item="{
+    cover: null,
+    title: section.name,
+    subtitle: section.title,
+    coverStyle: 'rectangle'
+}">
+                <template v-slot:badges>
+                    <BadgeUi class="mb-1 me-1 fw-light"
+                        :text="section.visible ? 'Visível' : 'Oculto'"
+                        :variant="section.visible ? 'info' : 'secondary'"
+                        :icon="section.visible ? 'eye' : 'eyeSlash'" />
+                    <BadgeUi class="mb-1 me-1 fw-light"
+                        :text="`Tipo de seção: ${terms.type['type_' + section.type]} `"
+                        variant="warning" />
+                </template>
 
-    <div class="row justify-content-center">
-        <ListItem v-for="section in sections?.data" :key="section.id" :item="{
-            cover: null,
-            title: section.name,
-            subtitle: section.title,
-            coverStyle: 'rectangle'
-        }">
-            <template v-slot:badges>
-                <BadgeUi class="mb-1 me-1 fw-light"
-                    :text="section.visible ? 'Visível' : 'Oculto'"
-                    :variant="section.visible ? 'info' : 'secondary'"
-                    :icon="section.visible ? 'eye' : 'eyeSlash'" />
-                <BadgeUi class="mb-1 me-1 fw-light"
-                    :text="`Tipo de seção: ${terms.type['type_' + section.type]} `"
-                    variant="warning" />
-            </template>
+                <template v-slot:actions>
+                    <ButtonUi icon="pencilSquare" variant="primary" size="sm" :to="$route('admin.sections.edit', {
+    section: section.id
+})" />
 
-            <template v-slot:actions>
-                <ButtonUi icon="pencilSquare" variant="primary" size="sm" :to="$route('admin.sections.edit', {
-                    section: section.id
-                })" />
-
-                <ButtonConfirmationUi v-if="section.can.delete" confirm-text="Excluir?" icon="trash"
-                    variant="danger" size="sm" class="ms-2" position="right" :data-action="$route('admin.sections.destroy',
-                    { section: section.id })" confirm-with-request
-                    request-method="delete" />
-            </template>
-        </ListItem>
-    </div>
-
-    <PaginationUi label="Paginação de seções" :pages="sections?.meta?.links" />
+                    <ButtonConfirmationUi v-if="section.can.delete"
+                        confirm-text="Excluir?" icon="trash" variant="danger" size="sm"
+                        class="ms-2" position="right" :data-action="$route('admin.sections.destroy',
+    { section: section.id })" confirm-with-request request-method="delete" />
+                </template>
+            </ListItem>
+        </div>
+    </ListGroup>
 
 </template>
 
 <script>
 
-import EmptyList from '../../../../Components/EmptyList.vue';
-import ListItem from './../../../../Components/ListItem.vue';
-import PaginationUi from '../../../../Components/PaginationUi.vue';
-import CardUi from '../../../../Components/Ui/CardUi.vue';
 import Layout from './../../../../Layouts/Panel.vue';
+import ListGroup from '../../../../Components/List/ListGroup.vue';
+import ListItem from './../../../../Components/List/ListItem.vue';
+import CardUi from '../../../../Components/Ui/CardUi.vue';
 import ButtonUi from '../../../../Components/Ui/ButtonUi.vue';
 import ButtonConfirmationUi from '../../../../Components/Ui/ButtonConfirmationUi.vue';
 import BadgeUi from '../../../../Components/Ui/BadgeUi.vue';
@@ -49,7 +47,7 @@ import BadgeUi from '../../../../Components/Ui/BadgeUi.vue';
 export default {
     layout: (h, page) => h(Layout, () => child),
     layout: Layout,
-    components: { EmptyList, ListItem, PaginationUi, CardUi, ButtonUi, ButtonConfirmationUi, BadgeUi },
+    components: { ListItem, CardUi, ButtonUi, ButtonConfirmationUi, BadgeUi, ListGroup },
 
     props: {
         sections: { type: Object, default: [] },
